todo exceptions

+++ src/main/java/com/pavel/webbet/service/impl/DoDeleteMatchCommand.java
@@ -1,6 +1,10 @@
 package com.pavel.webbet.service.impl;
 
 import com.pavel.webbet.constant.UrlConstant;
+import com.pavel.webbet.dao.IBetDao;
+import com.pavel.webbet.dao.IFootballMatchDao;
+import com.pavel.webbet.dao.factory.DaoFactory;
+import com.pavel.webbet.dao.factory.DaoType;
 import com.pavel.webbet.dao.mysql.MysqlDaoException;
 import com.pavel.webbet.dao.mysql.impl.FootballMatchDAO;
 import com.pavel.webbet.service.CommandException;
@@ -17,9 +21,15 @@ public class DoDeleteMatchCommand implements ICommand{
     @Override
     public String execute(HttpServletRequest request) throws CommandException {
         int id = Integer.parseInt(request.getParameter(PARAMETER_MATCH_TO_DELETE_ID));
-        FootballMatchDAO dao = FootballMatchDAO.getInstance();
+        IFootballMatchDao dao = DaoFactory.getDao(DaoType.MATCH);
+        IBetDao betDao = DaoFactory.getDao(DaoType.BET);
+
+
         try {
-            dao.deleteBean(id);
+            if (!betDao.containBetOnMatchId(id)){
+                dao.deleteBean(id);
+            }
+            else throw new CommandException("This match cannot be deleted: bets are done.");
         }
         catch (MysqlDaoException e){
             throw new CommandException(e.getMessage(), e);
